smart-library-ai/crawler/requirements.txt
# Web Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.8.0

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Database
pymysql>=1.0.0
sqlalchemy>=2.0.0

# Image Processing
Pillow>=9.0.0

# Utilities
python-dotenv>=0.20.0
tqdm>=4.64.0

# MinIO
minio>=7.1.0
